# docs/swagger.yaml
basePath: /
definitions:
...
  iku.IkuOpdResponse:
    properties:
      asal_iku:
        type: string
      created_at:
        type: string
      definisi_operasional:
        type: string
      iku_active:
        type: boolean
      indikator:
        type: string
      indikator_id:
        type: string
      is_hide:
        type: boolean
      jenis:
        type: string
      jenis_periode:
        type: string
      rumus_perhitungan:
        type: string
      sumber_data:
        type: string
      tahun_akhir:
        type: string
      tahun_awal:
        type: string
      target:
        items:
          $ref: '#/definitions/iku.TargetOpdResponse'
        type: array
    type: object
  iku.IkuUpdateActiveRequest:
    properties:
      indikator_id:
        type: string
      is_active:
        type: boolean
    type: object
  iku.TargetOpdResponse:
    properties:
      satuan:
        type: string
      tahun:
        type: string
      target:
        type: string
    type: object
...
  web.WebResponse:
    properties:
      code:
        type: integer
      data: {}
      error:
        type: string
      status:
        type: string
    type: object
info:
  contact:
    name: Tim Pengembang
  description: API Backend Kertas Kerja Kabupaten Madiun
  termsOfService: http://swagger.io/terms/
  title: EKAK Kabupaten Madiun API
  version: "1.0"
paths:
...
  /iku_renja_opd/penetapan/{kode_opd}/{tahun}:
    get:
      consumes:
      - application/json
      description: Mendapatkan semua IKU Renja Opd Penetapan berdasarkan kode OPD
        dan tahun.
      parameters:
      - description: Kode OPD
        example: '"1.01.1.01.0.00.01.0000"'
        in: path
        name: kode_opd
        required: true
        type: string
      - description: Tahun
        example: '"2025"'
        in: path
        name: tahun
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/iku.IkuOpdResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
      security:
      - BearerAuth: []
      summary: Find All IKU Renja Opd Penetapan
      tags:
      - IKU Renja Opd
  /iku_renja_opd/rankhir/{kode_opd}/{tahun}:
    get:
      consumes:
      - application/json
      description: Mendapatkan semua IKU Renja Opd Rankhir berdasarkan kode OPD dan
        tahun.
      parameters:
      - description: Kode OPD
        example: '"1.01.1.01.0.00.01.0000"'
        in: path
        name: kode_opd
        required: true
        type: string
      - description: Tahun
        example: '"2025"'
        in: path
        name: tahun
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/iku.IkuOpdResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
      security:
      - BearerAuth: []
      summary: Find All IKU Renja Opd Rankhir
      tags:
      - IKU Renja Opd
  /iku_renja_opd/ranwal/{kode_opd}/{tahun}:
    get:
      consumes:
      - application/json
      description: Mendapatkan semua IKU Renja Opd Ranwal berdasarkan kode OPD dan
        tahun.
      parameters:
      - description: Kode OPD
        example: '"1.01.1.01.0.00.01.0000"'
        in: path
        name: kode_opd
        required: true
        type: string
      - description: Tahun
        example: '"2025"'
        in: path
        name: tahun
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/iku.IkuOpdResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
      security:
      - BearerAuth: []
      summary: Find All IKU Renja Opd Ranwal
      tags:
      - IKU Renja Opd
  /indikator_utama/opd/{kode_opd}/{tahun_awal}/{tahun_akhir}/{jenis_periode}:
    get:
      consumes:
      - application/json
      description: Get IKU Renstra by kode OPD, tahun awal, tahun akhir, and jenis
        periode.
      parameters:
      - description: Kode OPD
        example: '"1.01.1.01.0.00.01.0000"'
        in: path
        name: kode_opd
        required: true
        type: string
      - description: Tahun Awal
        in: path
        name: tahun_awal
        required: true
        type: string
      - description: Tahun Akhir
        in: path
        name: tahun_akhir
        required: true
        type: string
      - description: Jenis Periode
        in: path
        name: jenis_periode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/iku.IkuOpdResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
      security:
      - BearerAuth: []
      summary: Get IKU Renstra OPD
      tags:
      - IKU Renstra
  /indikator_utama/opd/status/{kode_indikator}:
    put:
      consumes:
      - application/json
      description: Mengupdate status IKU Opd berdasarkan ID Indikator.
      parameters:
      - description: Kode Indikator
        in: path
        name: kode_indikator
        required: true
        type: string
      - description: Data untuk mengupdate status IKU
        in: body
        name: iku_update_active_request
        required: true
        schema:
          $ref: '#/definitions/iku.IkuUpdateActiveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
      security:
      - BearerAuth: []
      summary: Update IKU Opd Active
      tags:
      - IKU Renja Opd
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"